+++ server/CrowdfundingPlatform/CrowdfundingPlatform/src/main/java/com/example/CrowdfundingPlatform/Dto/CampaignDto.java
@@ -13,15 +13,15 @@ public class CampaignDto {
 
     public String description;
 
-    public long goalAmount;
+    public double goalAmount;
 
     public String deadline;
 
     public long userId;
 
-    public long  donatedAmount;
+    public double  donatedAmount;
 
-    public CampaignDto(String name, String description, long goalAmount, String deadline, long userId, long donatedAmount) {
+    public CampaignDto(String name, String description, double goalAmount, String deadline, long userId, double donatedAmount) {
         this.name = name;
         this.description = description;
         this.goalAmount = goalAmount;
