+++ server/CrowdfundingPlatform/CrowdfundingPlatform/src/main/java/com/example/CrowdfundingPlatform/Model/Campaign.java
@@ -22,15 +22,15 @@ public class Campaign {
 
     public String description;
 
-    public long goalAmount;
+    public double goalAmount;
 
     public String deadline;
 
     public long userId;
 
-    public long  donatedAmount;
+    public double  donatedAmount;
 
-    public Campaign(String name, String description, long goalAmount, String deadline, long userId, long donatedAmount) {
+    public Campaign(String name, String description, double goalAmount, String deadline, long userId, double donatedAmount) {
         this.name = name;
         this.description = description;
         this.goalAmount = goalAmount;
